scripts/requirements.txt
clickhouse-connect==0.7.0
psycopg2-binary==2.9.9